Categorical Data Classification Analysis

It performs the Categorical Data Classification Analysis from a data.frame of categorical data

CategoricalDataClassificationAnalysis(db, n_model, coordinates = "two")

Arguments

db

Database predictor and response variables. It must contain either character data or factors. The last column must contain the response variable

n_model

Number of individuals to build the map. The number of individuals in the test set is calculated by difference.

coordinates

Coordinates used in the discrimination. By default, the two polar coordinates will be used. If you only want to do it for one, you must indicate the name of the coordinate (theta or rho).

Value

List of calculated predictive parameters (relative confussion matrix, negative and positive predictive values, F1 score, Cohen's kappa and Matthew's correlation coefficients)